+++ slangpy/types/__init__.py
@@ -5,6 +5,7 @@ from .buffer import NDBuffer
 from .diffpair import DiffPair, diffPair, floatDiffPair
 from .randfloatarg import RandFloatArg, rand_float
 from .threadidarg import ThreadIdArg, thread_id
+from .callidarg import CallIdArg, call_id
 from .valueref import ValueRef, floatRef, intRef
 from .wanghasharg import WangHashArg, wang_hash
 from .tensor import Tensor